backend/app/requirements.txt
bcrypt==4.3.0
cryptography==44.0.0
fastapi==0.115.0
httpx==0.28.1
psycopg2-binary==2.9.10
pydantic==2.11.0
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
SQLAlchemy==2.0.36
uvicorn==0.34.0
